mysite/crimes/data/requirements.txt
requests
python-dotenv
django
pandas
seaborn
matplotlib